+++ kubernetes-tests/src/test/java/io/fabric8/kubernetes/client/mock/StatefulSetTest.java
@@ -30,7 +30,9 @@ import io.fabric8.kubernetes.api.model.apps.StatefulSetBuilder;
 import io.fabric8.kubernetes.api.model.apps.StatefulSetList;
 import io.fabric8.kubernetes.api.model.apps.StatefulSetListBuilder;
 import io.fabric8.kubernetes.client.KubernetesClient;
+import io.fabric8.kubernetes.client.KubernetesClientException;
 import io.fabric8.kubernetes.client.dsl.Deletable;
+import io.fabric8.kubernetes.client.dsl.TimeoutImageEditReplacePatchable;
 import io.fabric8.kubernetes.client.server.mock.EnableKubernetesMockClient;
 import io.fabric8.kubernetes.client.server.mock.KubernetesMockServer;
 import io.fabric8.kubernetes.client.utils.Utils;
@@ -49,6 +51,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @EnableKubernetesMockClient
@@ -59,11 +62,18 @@ public class StatefulSetTest {
 
   @Test
   public void testList() {
-    server.expect().withPath("/apis/apps/v1/namespaces/test/statefulsets").andReturn(200, new StatefulSetListBuilder().build())
+    server.expect()
+        .withPath("/apis/apps/v1/namespaces/test/statefulsets")
+        .andReturn(200, new StatefulSetListBuilder().build())
         .once();
-    server.expect().withPath("/apis/apps/v1/namespaces/ns1/statefulsets").andReturn(200, new StatefulSetListBuilder()
-        .addNewItem().and()
-        .addNewItem().and().build())
+    server.expect()
+        .withPath("/apis/apps/v1/namespaces/ns1/statefulsets")
+        .andReturn(200, new StatefulSetListBuilder()
+            .addNewItem()
+            .and()
+            .addNewItem()
+            .and()
+            .build())
         .once();
 
     StatefulSetList statefulSetList = client.apps().statefulSets().list();
@@ -77,9 +87,13 @@ public class StatefulSetTest {
 
   @Test
   public void testGet() {
-    server.expect().withPath("/apis/apps/v1/namespaces/test/statefulsets/repl1")
-        .andReturn(200, new StatefulSetBuilder().build()).once();
-    server.expect().withPath("/apis/apps/v1/namespaces/ns1/statefulsets/repl2").andReturn(200, new StatefulSetBuilder().build())
+    server.expect()
+        .withPath("/apis/apps/v1/namespaces/test/statefulsets/repl1")
+        .andReturn(200, new StatefulSetBuilder().build())
+        .once();
+    server.expect()
+        .withPath("/apis/apps/v1/namespaces/ns1/statefulsets/repl2")
+        .andReturn(200, new StatefulSetBuilder().build())
         .once();
 
     StatefulSet repl1 = client.apps().statefulSets().withName("repl1").get();
@@ -94,7 +108,8 @@ public class StatefulSetTest {
 
   @Test
   public void testDelete() {
-    server.expect().withPath("/apis/apps/v1/namespaces/test/statefulsets/repl1")
+    server.expect()
+        .withPath("/apis/apps/v1/namespaces/test/statefulsets/repl1")
         .andReturn(200, new StatefulSetBuilder().withNewMetadata()
             .withName("repl1")
             .withResourceVersion("1")
@@ -108,7 +123,8 @@ public class StatefulSetTest {
             .build())
         .once();
 
-    server.expect().withPath("/apis/apps/v1/namespaces/test/statefulsets/repl1")
+    server.expect()
+        .withPath("/apis/apps/v1/namespaces/test/statefulsets/repl1")
         .andReturn(200, new StatefulSetBuilder().withNewMetadata()
             .withName("repl1")
             .withResourceVersion("1")
@@ -122,7 +138,8 @@ public class StatefulSetTest {
             .build())
         .times(5);
 
-    server.expect().withPath("/apis/apps/v1/namespaces/ns1/statefulsets/repl2")
+    server.expect()
+        .withPath("/apis/apps/v1/namespaces/ns1/statefulsets/repl2")
         .andReturn(200, new StatefulSetBuilder().withNewMetadata()
             .withName("repl2")
             .withResourceVersion("1")
@@ -136,7 +153,8 @@ public class StatefulSetTest {
             .build())
         .once();
 
-    server.expect().withPath("/apis/apps/v1/namespaces/ns1/statefulsets/repl2")
+    server.expect()
+        .withPath("/apis/apps/v1/namespaces/ns1/statefulsets/repl2")
         .andReturn(200, new StatefulSetBuilder().withNewMetadata()
             .withName("repl2")
             .withResourceVersion("1")
@@ -163,7 +181,10 @@ public class StatefulSetTest {
   @Test
   public void testDeleteLoadedResource() {
     StatefulSet response = client.apps().statefulSets().load(getClass().getResourceAsStream("/test-statefulset.yml")).get();
-    server.expect().delete().withPath("/apis/apps/v1/namespaces/test/statefulsets/example").andReturn(200, response)
+    server.expect()
+        .delete()
+        .withPath("/apis/apps/v1/namespaces/test/statefulsets/example")
+        .andReturn(200, response)
         .once();
 
     Deletable items = client.load(getClass().getResourceAsStream("/test-statefulset.yml"));
@@ -172,18 +193,21 @@ public class StatefulSetTest {
 
   @Test
   public void testScale() {
-    server.expect().withPath("/apis/apps/v1/namespaces/test/statefulsets/repl1").andReturn(200, new StatefulSetBuilder()
-        .withNewMetadata()
-        .withName("repl1")
-        .withResourceVersion("1")
-        .endMetadata()
-        .withNewSpec()
-        .withReplicas(5)
-        .endSpec()
-        .withNewStatus()
-        .withReplicas(1)
-        .endStatus()
-        .build()).always();
+    server.expect()
+        .withPath("/apis/apps/v1/namespaces/test/statefulsets/repl1")
+        .andReturn(200, new StatefulSetBuilder()
+            .withNewMetadata()
+            .withName("repl1")
+            .withResourceVersion("1")
+            .endMetadata()
+            .withNewSpec()
+            .withReplicas(5)
+            .endSpec()
+            .withNewStatus()
+            .withReplicas(1)
+            .endStatus()
+            .build())
+        .always();
 
     StatefulSet repl = client.apps().statefulSets().withName("repl1").scale(5);
     assertNotNull(repl);
@@ -194,18 +218,21 @@ public class StatefulSetTest {
 
   @Test
   public void testScaleAndWait() {
-    server.expect().withPath("/apis/apps/v1/namespaces/test/statefulsets/repl1").andReturn(200, new StatefulSetBuilder()
-        .withNewMetadata()
-        .withName("repl1")
-        .withResourceVersion("1")
-        .endMetadata()
-        .withNewSpec()
-        .withReplicas(5)
-        .endSpec()
-        .withNewStatus()
-        .withReplicas(1)
-        .endStatus()
-        .build()).once();
+    server.expect()
+        .withPath("/apis/apps/v1/namespaces/test/statefulsets/repl1")
+        .andReturn(200, new StatefulSetBuilder()
+            .withNewMetadata()
+            .withName("repl1")
+            .withResourceVersion("1")
+            .endMetadata()
+            .withNewSpec()
+            .withReplicas(5)
+            .endSpec()
+            .withNewStatus()
+            .withReplicas(1)
+            .endStatus()
+            .build())
+        .once();
 
     StatefulSet scaled = new StatefulSetBuilder()
         .withNewMetadata()
@@ -246,7 +273,9 @@ public class StatefulSetTest {
         .withNewSpec()
         .withReplicas(1)
         .withNewTemplate()
-        .withNewMetadata().withLabels(new HashMap<String, String>()).endMetadata()
+        .withNewMetadata()
+        .withLabels(new HashMap<String, String>())
+        .endMetadata()
         .withNewSpec()
         .addNewContainer()
         .withImage("img1")
@@ -254,17 +283,27 @@ public class StatefulSetTest {
         .endSpec()
         .endTemplate()
         .endSpec()
-        .withNewStatus().withReplicas(1).endStatus()
+        .withNewStatus()
+        .withReplicas(1)
+        .endStatus()
         .build();
 
     server.expect().withPath("/apis/apps/v1/namespaces/test/statefulsets/repl1").andReturn(200, repl1).once();
     server.expect().put().withPath("/apis/apps/v1/namespaces/test/statefulsets/repl1").andReturn(200, repl1).once();
-    server.expect().get().withPath("/apis/apps/v1/namespaces/test/statefulsets")
-        .andReturn(200, new StatefulSetListBuilder().withItems(repl1).build()).once();
+    server.expect()
+        .get()
+        .withPath("/apis/apps/v1/namespaces/test/statefulsets")
+        .andReturn(200, new StatefulSetListBuilder().withItems(repl1).build())
+        .once();
     server.expect().post().withPath("/apis/apps/v1/namespaces/test/statefulsets").andReturn(201, repl1).once();
-    server.expect().withPath("/apis/apps/v1/namespaces/test/pods").andReturn(200, new KubernetesListBuilder().build()).once();
+    server.expect()
+        .withPath("/apis/apps/v1/namespaces/test/pods")
+        .andReturn(200, new KubernetesListBuilder().build())
+        .once();
 
-    repl1 = client.apps().statefulSets().withName("repl1")
+    repl1 = client.apps()
+        .statefulSets()
+        .withName("repl1")
         .rolling()
         .withTimeout(5, TimeUnit.MINUTES)
         .updateImage("");
@@ -276,19 +315,34 @@ public class StatefulSetTest {
   void testRolloutUpdateSingleImage() throws InterruptedException {
     // Given
     String imageToUpdate = "nginx:latest";
-    server.expect().get().withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
-        .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder().build()).times(3);
-    server.expect().patch().withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
+    server.expect()
+        .get()
+        .withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
+        .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder().build())
+        .times(3);
+    server.expect()
+        .patch()
+        .withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
         .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder()
-            .editSpec().editTemplate().editSpec().editContainer(0)
+            .editSpec()
+            .editTemplate()
+            .editSpec()
+            .editContainer(0)
             .withImage(imageToUpdate)
-            .endContainer().endSpec().endTemplate().endSpec()
+            .endContainer()
+            .endSpec()
+            .endTemplate()
+            .endSpec()
             .build())
         .once();
 
     // When
-    StatefulSet statefulSet = client.apps().statefulSets().inNamespace("ns1").withName("statefulset1")
-        .rolling().updateImage(imageToUpdate);
+    StatefulSet statefulSet = client.apps()
+        .statefulSets()
+        .inNamespace("ns1")
+        .withName("statefulset1")
+        .rolling()
+        .updateImage(imageToUpdate);
 
     // Then
     RecordedRequest recordedRequest = server.getLastRequest();
@@ -303,19 +357,34 @@ public class StatefulSetTest {
   void testRolloutUpdateImage() throws InterruptedException {
     // Given
     Map<String, String> containerToImageMap = Collections.singletonMap("nginx", "nginx:latest");
-    server.expect().get().withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
-        .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder().build()).times(3);
-    server.expect().patch().withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
+    server.expect()
+        .get()
+        .withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
+        .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder().build())
+        .times(3);
+    server.expect()
+        .patch()
+        .withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
         .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder()
-            .editSpec().editTemplate().editSpec().editContainer(0)
+            .editSpec()
+            .editTemplate()
+            .editSpec()
+            .editContainer(0)
             .withImage(containerToImageMap.get("nginx"))
-            .endContainer().endSpec().endTemplate().endSpec()
+            .endContainer()
+            .endSpec()
+            .endTemplate()
+            .endSpec()
             .build())
         .once();
 
     // When
-    StatefulSet deployment = client.apps().statefulSets().inNamespace("ns1").withName("statefulset1")
-        .rolling().updateImage(containerToImageMap);
+    StatefulSet deployment = client.apps()
+        .statefulSets()
+        .inNamespace("ns1")
+        .withName("statefulset1")
+        .rolling()
+        .updateImage(containerToImageMap);
 
     // Then
     assertNotNull(deployment);
@@ -325,55 +394,79 @@ public class StatefulSetTest {
   }
 
   @Test
-  @DisplayName("Should pause resource")
-  void testRolloutPause() throws InterruptedException {
+  @DisplayName("Should not pause resource")
+  void testRolloutPauseUnsupported() throws InterruptedException {
     // Given
-    server.expect().get().withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
-        .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder().build()).times(3);
-    server.expect().patch().withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
-        .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder().build()).once();
+    server.expect()
+        .get()
+        .withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
+        .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder().build())
+        .times(3);
+    server.expect()
+        .patch()
+        .withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
+        .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder().build())
+        .once();
 
     // When
-    StatefulSet deployment = client.apps().statefulSets().inNamespace("ns1").withName("statefulset1")
-        .rolling().pause();
+    TimeoutImageEditReplacePatchable<StatefulSet> rolling = client.apps()
+        .statefulSets()
+        .inNamespace("ns1")
+        .withName("statefulset1")
+        .rolling();
 
     // Then
-    RecordedRequest recordedRequest = server.getLastRequest();
-    assertEquals("PATCH", recordedRequest.getMethod());
-    assertEquals("{\"spec\":{\"paused\":true}}", recordedRequest.getBody().readUtf8());
+    assertThrows(KubernetesClientException.class, () -> rolling.pause());
   }
 
   @Test
-  @DisplayName("Should resume rollout")
-  void testRolloutResume() throws InterruptedException {
+  @DisplayName("Should not resume rollout")
+  void testRolloutResumeUnsupported() throws InterruptedException {
     // Given
-    server.expect().get().withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
-        .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder().build()).times(3);
-    server.expect().patch().withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
-        .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder().build()).once();
+    server.expect()
+        .get()
+        .withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
+        .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder().build())
+        .times(3);
+    server.expect()
+        .patch()
+        .withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
+        .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder().build())
+        .once();
 
     // When
-    StatefulSet deployment = client.apps().statefulSets().inNamespace("ns1").withName("statefulset1")
-        .rolling().resume();
+    TimeoutImageEditReplacePatchable<StatefulSet> rolling = client.apps()
+        .statefulSets()
+        .inNamespace("ns1")
+        .withName("statefulset1")
+        .rolling();
 
     // Then
-    RecordedRequest recordedRequest = server.getLastRequest();
-    assertEquals("PATCH", recordedRequest.getMethod());
-    assertEquals("{\"spec\":{\"paused\":null}}", recordedRequest.getBody().readUtf8());
+    assertThrows(KubernetesClientException.class, () -> rolling.resume());
   }
 
   @Test
   @DisplayName("Should restart rollout")
   void testRolloutRestart() throws InterruptedException {
     // Given
-    server.expect().get().withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
-        .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder().build()).times(3);
-    server.expect().patch().withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
-        .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder().build()).once();
+    server.expect()
+        .get()
+        .withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
+        .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder().build())
+        .times(3);
+    server.expect()
+        .patch()
+        .withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
+        .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder().build())
+        .once();
 
     // When
-    StatefulSet deployment = client.apps().statefulSets().inNamespace("ns1").withName("statefulset1")
-        .rolling().restart();
+    StatefulSet deployment = client.apps()
+        .statefulSets()
+        .inNamespace("ns1")
+        .withName("statefulset1")
+        .rolling()
+        .restart();
 
     // Then
     RecordedRequest recordedRequest = server.getLastRequest();
@@ -395,7 +488,9 @@ public class StatefulSetTest {
         .withNewStatefulSetData()
         .withNewSpec()
         .withReplicas(0)
-        .withNewSelector().addToMatchLabels("app", "nginx").endSelector()
+        .withNewSelector()
+        .addToMatchLabels("app", "nginx")
+        .endSelector()
         .withNewTemplate()
         .withNewMetadata()
         .addToAnnotations("kubectl.kubernetes.io/restartedAt", "2020-06-08T11:52:50.022")
@@ -406,7 +501,9 @@ public class StatefulSetTest {
         .addNewContainer()
         .withName("nginx")
         .withImage("nginx:perl")
-        .addNewPort().withContainerPort(80).endPort()
+        .addNewPort()
+        .withContainerPort(80)
+        .endPort()
         .endContainer()
         .endSpec()
         .endTemplate()
@@ -422,7 +519,9 @@ public class StatefulSetTest {
         .withNewStatefulSetData()
         .withNewSpec()
         .withReplicas(1)
-        .withNewSelector().addToMatchLabels("app", "nginx").endSelector()
+        .withNewSelector()
+        .addToMatchLabels("app", "nginx")
+        .endSelector()
         .withNewTemplate()
         .withNewMetadata()
         .addToAnnotations("kubectl.kubernetes.io/restartedAt", "2020-06-08T11:52:50.022")
@@ -433,7 +532,9 @@ public class StatefulSetTest {
         .addNewContainer()
         .withName("nginx")
         .withImage("nginx:1.19")
-        .addNewPort().withContainerPort(80).endPort()
+        .addNewPort()
+        .withContainerPort(80)
+        .endPort()
         .endContainer()
         .endSpec()
         .endTemplate()
@@ -441,19 +542,30 @@ public class StatefulSetTest {
         .endStatefulSetData()
         .build();
 
-    server.expect().get()
+    server.expect()
+        .get()
         .withPath("/apis/apps/v1/namespaces/ns1/controllerrevisions?labelSelector=" + Utils.toUrlEncoded("app=nginx"))
         .andReturn(HttpURLConnection.HTTP_OK,
             new ControllerRevisionListBuilder().withItems(controllerRevision1, controllerRevision2).build())
         .once();
-    server.expect().get().withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
-        .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder().build()).times(3);
-    server.expect().patch().withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
-        .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder().build()).once();
+    server.expect()
+        .get()
+        .withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
+        .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder().build())
+        .times(3);
+    server.expect()
+        .patch()
+        .withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
+        .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder().build())
+        .once();
 
     // When
-    StatefulSet deployment = client.apps().statefulSets().inNamespace("ns1").withName("statefulset1")
-        .rolling().undo();
+    StatefulSet deployment = client.apps()
+        .statefulSets()
+        .inNamespace("ns1")
+        .withName("statefulset1")
+        .rolling()
+        .undo();
 
     // Then
     RecordedRequest recordedRequest = server.getLastRequest();
@@ -467,14 +579,20 @@ public class StatefulSetTest {
     // Given
     Pod jobPod = createPod();
 
-    server.expect().get().withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
+    server.expect()
+        .get()
+        .withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
         .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder().build())
         .always();
 
-    server.expect().get().withPath("/api/v1/namespaces/ns1/pods?labelSelector=app%3Dnginx")
+    server.expect()
+        .get()
+        .withPath("/api/v1/namespaces/ns1/pods?labelSelector=app%3Dnginx")
         .andReturn(HttpURLConnection.HTTP_OK, new PodListBuilder().withItems(jobPod).build())
         .once();
-    server.expect().get().withPath("/api/v1/namespaces/ns1/pods/ss-hk9nf/log?pretty=false")
+    server.expect()
+        .get()
+        .withPath("/api/v1/namespaces/ns1/pods/ss-hk9nf/log?pretty=false")
         .andReturn(HttpURLConnection.HTTP_OK, "hello")
         .once();
 
@@ -492,14 +610,20 @@ public class StatefulSetTest {
     // Given
     Pod jobPod = createPod();
 
-    server.expect().get().withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
+    server.expect()
+        .get()
+        .withPath("/apis/apps/v1/namespaces/ns1/statefulsets/statefulset1")
         .andReturn(HttpURLConnection.HTTP_OK, createStatefulSetBuilder().build())
         .always();
 
-    server.expect().get().withPath("/api/v1/namespaces/ns1/pods?labelSelector=app%3Dnginx")
+    server.expect()
+        .get()
+        .withPath("/api/v1/namespaces/ns1/pods?labelSelector=app%3Dnginx")
         .andReturn(HttpURLConnection.HTTP_OK, new PodListBuilder().withItems(jobPod).build())
         .once();
-    server.expect().get().withPath("/api/v1/namespaces/ns1/pods/ss-hk9nf/log?pretty=false&container=c1")
+    server.expect()
+        .get()
+        .withPath("/api/v1/namespaces/ns1/pods/ss-hk9nf/log?pretty=false&container=c1")
         .andReturn(HttpURLConnection.HTTP_OK, "hello")
         .once();
 
@@ -521,7 +645,8 @@ public class StatefulSetTest {
             .withName("pi")
             .withUid("3Dc4c8746c-94fd-47a7-ac01-11047c0323b4")
             .build())
-        .withName("ss-hk9nf").addToLabels("controller-uid", "3Dc4c8746c-94fd-47a7-ac01-11047c0323b4")
+        .withName("ss-hk9nf")
+        .addToLabels("controller-uid", "3Dc4c8746c-94fd-47a7-ac01-11047c0323b4")
         .endMetadata()
         .build();
   }
@@ -540,12 +665,16 @@ public class StatefulSetTest {
         .addToMatchLabels("app", "nginx")
         .endSelector()
         .withNewTemplate()
-        .withNewMetadata().addToLabels("app", "nginx").endMetadata()
+        .withNewMetadata()
+        .addToLabels("app", "nginx")
+        .endMetadata()
         .withNewSpec()
         .addNewContainer()
         .withName("nginx")
         .withImage("nginx:1.7.9")
-        .addNewPort().withContainerPort(80).endPort()
+        .addNewPort()
+        .withContainerPort(80)
+        .endPort()
         .endContainer()
         .endSpec()
         .endTemplate()